PDF_Summarizer_Gemini/requirements.txt
google-generativeai
gradio
PyPDF2
python-dotenv
tqdm
reportlab